+++ src/main/java/com/ngo/fundraiser/repository/DonorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DonorRepository extends JpaRepository<Donor, Long> {
+public interface DonorRepository extends JpaRepository<Donor, Integer> {
 }
